Almost Finished

+++ Airport.cpp
@@ -1,6 +1,6 @@
 #include "Airport.h"
 
-Airport::Airport(string code, string name, string city, string country, float latitude, float longitude) {
+Airport::Airport(string code, string name, string city, string country, double latitude, double longitude) {
     this->code=code;
     this->name=name;
     this->city=city;
