+++ Airport.h
@@ -10,11 +10,11 @@ class Airport {
     string name;
     string city;
     string country;
-    float latitude;
-    float longitude;
+    double latitude;
+    double longitude;
 public:
     Airport();
-    Airport(string code, string name, string city, string country, float latitude, float longitude);
+    Airport(string code, string name, string city, string country, double latitude, double longitude);
     string getCode();
     string getName();
     string getCity();
